validate with variables, need to add downloader logic in preset
