+++ pyproject.toml
@@ -13,6 +13,7 @@ disable = [
     "R0903", # too-few-public-methods
     "R0801", # similar lines
     "R0913", # Too many arguments
+    "R0901", # too-many-ancestors
     "W0511", # TODO
 ]
